requirements.txt
Django==3.1.7
djangorestframework==3.12.2
django-cors-headers==3.7.0
psycopg2
pyJWT==2.0.1
Faker==6.6.1